requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
psycopg2-binary==2.9.7
djongo==1.3.6
pymongo==3.12.3
lxml==4.9.3
xmltodict==0.13.0
django-filter==23.3
django-extensions==3.2.3
celery==5.3.4
redis==5.0.1
Pillow==10.1.0
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
